refactor: update import paths for teas and restore teas data structure
