+++ courses/backend/advanced-javascript/week4-assigment/4.js
@@ -1,4 +1,4 @@
-import { teas } from "../teas.js";
+import { teas } from "./teas.js";
 import { Tea } from "./exercise1.js";
 import { OrderItem, Order } from "./exercise2.js";
 class Customer {
